Edit account and car info (#35)
* account list database edit!

* account list database edit!

* account list database edit - BUG VERSION

* Accounts fixed

* Connect CarWindow.py to MainWindow.py

* Rename CarWindow.py to CarMgmtWindow.py + Rename Settings.py to AccountMgmtWindow.py

---------
